common action teamsToRemove
