+++ app/actions/remote/entry/notification.ts
@@ -11,13 +11,13 @@ import DatabaseManager from '@database/manager';
 import {queryChannelsById, queryDefaultChannelForTeam, queryMyChannel} from '@queries/servers/channel';
 import {prepareModels} from '@queries/servers/entry';
 import {queryCommonSystemValues, queryCurrentTeamId, queryWebSocketLastDisconnected, setCurrentTeamAndChannelId} from '@queries/servers/system';
-import {deleteMyTeams, queryMyTeamById, queryMyTeamsById, queryTeamsById} from '@queries/servers/team';
+import {queryMyTeamById} from '@queries/servers/team';
 import {queryCurrentUser} from '@queries/servers/user';
 import EphemeralStore from '@store/ephemeral_store';
 import {isTablet} from '@utils/helpers';
 import {emitNotificationError} from '@utils/notification';
 
-import {AppEntryData, AppEntryError, deferredAppEntryActions, fetchAppEntryData, syncOtherServers} from './common';
+import {AppEntryData, AppEntryError, deferredAppEntryActions, fetchAppEntryData, syncOtherServers, teamsToRemove} from './common';
 
 export const pushNotificationEntry = async (serverUrl: string, notification: NotificationWithData) => {
     const operator = DatabaseManager.serverDatabases[serverUrl]?.operator;
@@ -117,16 +117,7 @@ export const pushNotificationEntry = async (serverUrl: string, notification: Not
         emitNotificationError('Channel');
     }
 
-    let removeTeams;
-    if (removeTeamIds?.length) {
-        // Immediately delete myTeams so that the UI renders only teams the user is a member of.
-        const removeMyTeams = await queryMyTeamsById(database, removeTeamIds);
-        if (removeMyTeams?.length) {
-            await deleteMyTeams(operator, removeMyTeams);
-            const ids = removeMyTeams.map((m) => m.id);
-            removeTeams = await queryTeamsById(database, ids);
-        }
-    }
+    const removeTeams = await teamsToRemove(serverUrl, removeTeamIds);
 
     let removeChannels;
     if (removeChannelIds?.length) {
